--- rql/search/config/const.go
@@ -1,8 +1,0 @@
-package config
-
-const (
-	singular = "RQL config query"
-	plural   = "RQL config queries"
-)
-
-var Suffix = []string{"search", "config"}
