+++ rql/search/const.go
@@ -1,0 +1,8 @@
+package search
+
+const (
+	singular = "RQL not saved query"
+	plural   = "RQL not saved queries"
+)
+
+var Suffix = []string{"search"}
