--- rql/search/event/const.go
@@ -1,8 +1,0 @@
-package event
-
-const (
-	singular = "RQL event query"
-	plural   = "RQL event queries"
-)
-
-var Suffix = []string{"search", "event"}
