+++ rql/search/funcs.go
@@ -1,13 +1,15 @@
 package search
 
 import (
+	"strings"
+
 	pc "github.com/paloaltonetworks/prisma-cloud-go"
 	"github.com/paloaltonetworks/prisma-cloud-go/rql/history"
 )
 
 // Identify returns the ID for the given account group.
 func Identify(c pc.PrismaCloudClient, query string) (string, error) {
-	c.Log(pc.LogAction, "(get) id for %s: %s", "RQL search not saved", query)
+	c.Log(pc.LogAction, "(get) id for %s: %s", Singular, query)
 
 	listing, err := history.List(c, history.Recent, 0)
 	if err != nil {
@@ -21,4 +23,23 @@ func Identify(c pc.PrismaCloudClient, query string) (string, error) {
 	}
 
 	return "", pc.ObjectNotFoundError
+}
+
+func Create(c pc.PrismaCloudClient, query history.Query) error {
+	var (
+		logMsg strings.Builder
+	)
+
+	logMsg.Grow(30)
+	logMsg.WriteString("(create)")
+	logMsg.WriteString(singular)
+
+	c.Log(pc.LogAction, logMsg.String())
+
+	path := make([]string, 0, len(Suffix)+1)
+	path = append(path, Suffix...)
+	path = append(path, query.SearchType)
+
+	_, err := c.Communicate("POST", path, nil, query, nil)
+	return err
 }
